benchstat: rearrange central data structure
Currently, benchstat has a cascade of different data structures for
organizing the benchmark data in different ways. This makes it very
difficult to extend and results in a lot of repeated effort between
the 1, 2, and "many" cases.

Replace this with a single, unopinionated "Collection" structure that
simply maps from (config, benchmark, unit) to Benchstat. We read all
input files in to this structure at the beginning of main and the
various cases traverse the key space in whatever way is most
convenient.